requirements.txt
h5py==2.10.0
imutils==0.5.4
Keras==2.3.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.7.0.72
Pillow==8.4.0
PyYAML==6.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
tensorflow==1.14.0
